backend/requirements.txt
beautifulsoup4==4.12.3
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
requests==2.32.3